libcamera: software_isp: Pass color lookup tables as floats
The color lookup tables are passed from stats processing to debayering
for direct use as 8-bit color output values.  Let's pass the values as
0.0..1.0 floats to make the gains color bit width independent.

Completes software ISP TODO #4.

+++ include/libcamera/internal/software_isp/debayer_params.h
@@ -19,7 +19,7 @@ struct DebayerParams {
 	static constexpr unsigned int kRGBLookupSize = 256;
 	static constexpr float kGamma = 0.5;
 
-	using ColorLookupTable = std::array<uint8_t, kRGBLookupSize>;
+	using ColorLookupTable = std::array<float, kRGBLookupSize>;
 
 	ColorLookupTable red;
 	ColorLookupTable green;
